Use the implicit flow to avoid sharing the client secret
Unfortunately, it appears that Spotify requires that we provide a client
secret when trying to obtain an access token.  This means that we can't
use the more secure authorization_code flow with PKCE.

I don't really know enought about this topic to know if this will become
a problem or not.